requirements.txt
flask
paddleocr
paddlepaddle
pillow
numpy
